utl: ensure utl::metric correctly handles large integers

+++ src/utl/src/LoggerCommon.h
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Copyright (c) 2022-2025, The OpenROAD Authors
 
+#include <cstdint>
 #include <string>
 
 #include "utl/Logger.h"
@@ -15,7 +16,7 @@ void critical(utl::ToolId tool, int id, const char* msg);
 void open_metrics(const char* metrics_filename);
 void close_metrics(const char* metrics_filename);
 void metric(const char* metric, const char* value);
-void metric_integer(const char* metric, const int value);
+void metric_integer(const char* metric, const int64_t value);
 void metric_float(const char* metric, const double value);
 void metric_float(const char* metric, const char* value);
 void set_metrics_stage(const char* fmt);
